+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/node/cluster/Cluster.java
@@ -130,7 +130,7 @@ public class Cluster implements AutoCloseable {
     }
   }
 
-  private void startNode(final RunnableNode node) {
+  public void startNode(final RunnableNode node) {
     this.startNode(node, false);
   }
