+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/node/configuration/BesuNodeConfiguration.java
@@ -50,7 +50,7 @@ public class BesuNodeConfiguration {
   private final List<String> extraCLIOptions;
   private final List<String> staticNodes;
   private final Optional<PrivacyParameters> privacyParameters;
-  private final Optional<String> runCommand;
+  private final List<String> runCommand;
 
   BesuNodeConfiguration(
       final String name,
@@ -74,7 +74,7 @@ public class BesuNodeConfiguration {
       final List<String> extraCLIOptions,
       final List<String> staticNodes,
       final Optional<PrivacyParameters> privacyParameters,
-      final Optional<String> runCommand) {
+      final List<String> runCommand) {
     this.name = name;
     this.miningParameters = miningParameters;
     this.jsonRpcConfiguration = jsonRpcConfiguration;
@@ -183,7 +183,7 @@ public class BesuNodeConfiguration {
     return privacyParameters;
   }
 
-  public Optional<String> getRunCommand() {
+  public List<String> getRunCommand() {
     return runCommand;
   }
 }
